слегка переделаны методы categories

+++ main-service/main-service-admin/src/main/java/ru/yandex/practicum/ewm/categories/service/CategoryServiceImpl.java
@@ -6,23 +6,28 @@ import org.springframework.stereotype.Service;
 
 import ru.yandex.practicum.ewm.dto.CategoryDto.CategoryDto;
 import ru.yandex.practicum.ewm.dto.CategoryDto.NewCategoryDto;
+import ru.yandex.practicum.ewm.exception.exceptionType.ConflictException;
 import ru.yandex.practicum.ewm.exception.exceptionType.NotFoundException;
-import ru.yandex.practicum.ewm.exception.exceptionType.ValidationException;
 import ru.yandex.practicum.ewm.mapper.CategoryMapper;
 import ru.yandex.practicum.ewm.model.Category;
 import ru.yandex.practicum.ewm.repository.CategoryRepository;
+import ru.yandex.practicum.ewm.repository.EventRepository;
 
 
 @Service
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Override
     public CategoryDto addCategory(NewCategoryDto dto) {
         if (categoryRepository.existsByName(dto.getName())) {
-            throw new ValidationException("Категория с таким именем уже существует");
+            throw new ConflictException(
+                    "Категория с именем '" + dto.getName() + "' уже существует"
+            );
         }
+
         Category category = CategoryMapper.toEntity(dto);
         Category saved = categoryRepository.save(category);
         return CategoryMapper.toDto(saved);
@@ -32,21 +37,30 @@ public class CategoryServiceImpl implements CategoryService {
     public CategoryDto updateCategory(Long catId, CategoryDto dto) {
         Category category = categoryRepository.findById(catId)
                 .orElseThrow(() ->
-                        new NotFoundException("Категория не найдена"));
-        if (categoryRepository.existsByName(dto.getName())) {
-            throw new ValidationException("Категория с таким именем уже существует");
+                        new NotFoundException("Категория с именем =" + catId + " не найден")
+                );
+
+        if (categoryRepository.existsByNameAndIdNot(dto.getName(), catId)) {
+            throw new ConflictException(
+                    "Категория с именем '" + dto.getName() + "' уже существует"
+            );
         }
+
         category.setName(dto.getName());
         Category updated = categoryRepository.save(category);
         return CategoryMapper.toDto(updated);
-
     }
 
     @Override
     public void deleteCategory(Long catId) {
         if (!categoryRepository.existsById(catId)) {
-            throw new NotFoundException("Категория не найдена");
+            throw new NotFoundException("Категория с именем" + catId + " не найден");
         }
+
+        if (eventRepository.existsByCategoryId(catId)) {
+            throw new ConflictException("Данная категория не пустая");
+        }
+
         categoryRepository.deleteById(catId);
     }
